Create feature spec helper file
