+++ spec/features/users_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe "Tests about users", type: :request do
+RSpec.describe "Tests about users", type: :feature do
   before(:each) do
     stub_civicrm
     @user = FactoryGirl.create(:user)
