+++ spec/rails_helper.rb
@@ -70,11 +70,3 @@ def login(user)
     }
   }
 end
-
-# requires capybara
-def sign_in_user(user)
-  visit "/login"
-  fill_in "Email", with: user.email
-  fill_in "Password", with: user.password
-  click_button "Sign in"
-end
