+++ spec/support/feature_helpers.rb
@@ -1,0 +1,8 @@
+module FeatureHelpers
+  def sign_in_user(user)
+    visit "/login"
+    fill_in "Email", with: user.email
+    fill_in "Password", with: user.password
+    click_button "Sign in"
+  end
+end
